+++ .github/workflows/go.yml
@@ -34,7 +34,7 @@ jobs:
       run: go build -v .
 
     - name: Test
-      run: go test -v -race -covermode atomic -coverprofile=profile.cov ./...
+      run: go test -v -race -covermode atomic -coverprofile=profile.cov -tags=unit ./...
 
     - name: Send coverage
       env:
