+++ cmd/linter.go
@@ -1,10 +1,7 @@
 package main
 
 import (
-	"flag"
 	"io"
-	"log"
-	"os"
 	"path/filepath"
 	"text/template"
 
@@ -46,19 +43,4 @@ func run(wr io.Writer, opt options) error {
 	}
 
 	return nil
-}
-
-func main() {
-	dir := flag.String("d", ".", "Specifies the directory you want to use to lint the CODEOWNERS file")
-	format := flag.String("f", "", "Specifies the format you want to return lint results")
-	flag.Parse()
-
-	opt := options{
-		directory: *dir,
-		format:    *format,
-	}
-
-	if err := run(os.Stdout, opt); err != nil {
-		log.Fatal(err)
-	}
 }
